6sizse

+++ 0x00-hello_world/6-size.c
@@ -8,10 +8,12 @@
 
 int main(void)
 {
-	printf("Size of a char: %lu bytes(s)\n", sizeof(char));
-	printf("Size of an int: %lu bytes(s)\n", sizeof(int));
-	printf("Size of a long int: %lu bytes(s)\n", sizeof(long int));
-	printf("Size of a long long int: %lu bytes(s)\n", sizeof(long long int));
-	printf("Size of a float: %lu bytes(s)\n", sizeof(float));
+	printf("Size of a char: %lu bytes(s)\n", (long int)sizeof(char));
+	printf("Size of an int: %lu bytes(s)\n", (long int)sizeof(int));
+	printf("Size of a long int: %lu bytes(s)\n", (long int)sizeof(long int));
+	printf("Size of a long long int: %lu bytes(s)\n",
+		(long int)sizeof(long long int));
+	printf("Size of a float: %lu bytes(s)\n", (long int)sizeof(float));
 	return (0);
 }
+
